Rename leo commands. delete -> destroy read -> status
